Update Travis and testing demo
