+++ .travis.yml
@@ -18,8 +18,13 @@ before_install:
 
 install:
   - pip install -e .
+  - cd dataset
+  - bash download.sh
+  - cd ..
 
 script:
   - python -c "import keras.backend"
   - sed -i -e 's/"backend":[[:space:]]*"[^"]*/"backend":\ "'$KERAS_BACKEND'/g' ~/.keras/keras.json;
   - echo -e "Running tests with the following config:\n$(cat ~/.keras/keras.json)"
+  - cd pria_lifechem/models
+  - python test_demo.py
